Move TimeSettings and TimeSettingsValidator into calendar package
closes #4976
