+++ src/main/java/org/synyx/urlaubsverwaltung/application/application/ApplicationMailService.java
@@ -4,7 +4,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import org.synyx.urlaubsverwaltung.absence.TimeSettings;
 import org.synyx.urlaubsverwaltung.application.comment.ApplicationComment;
 import org.synyx.urlaubsverwaltung.calendar.AbsenceTimeConfiguration;
 import org.synyx.urlaubsverwaltung.calendar.CalendarAbsence;
@@ -1004,8 +1003,7 @@ class ApplicationMailService {
     }
 
     private AbsenceTimeConfiguration getAbsenceTimeConfiguration() {
-        final TimeSettings timeSettings = settingsService.getSettings().getTimeSettings();
-        return new AbsenceTimeConfiguration(timeSettings);
+        return new AbsenceTimeConfiguration(settingsService.getSettings().getTimeSettings());
     }
 
     private static <T, R> Function<T, R> cachedByKey(Function<T, R> function) {
