+++ src/main/java/org/synyx/urlaubsverwaltung/calendar/CalendarAbsenceService.java
@@ -3,7 +3,6 @@ package org.synyx.urlaubsverwaltung.calendar;
 import org.apache.commons.collections4.ListUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.synyx.urlaubsverwaltung.absence.TimeSettings;
 import org.synyx.urlaubsverwaltung.application.application.Application;
 import org.synyx.urlaubsverwaltung.application.application.ApplicationService;
 import org.synyx.urlaubsverwaltung.application.application.ApplicationStatus;
