+++ src/test/java/org/synyx/urlaubsverwaltung/calendar/AbsenceTimeConfigurationTest.java
@@ -2,7 +2,6 @@ package org.synyx.urlaubsverwaltung.calendar;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.synyx.urlaubsverwaltung.absence.TimeSettings;
 
 import java.time.LocalTime;
