+++ src/test/java/org/synyx/urlaubsverwaltung/calendar/TimeSettingsTest.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.absence;
+package org.synyx.urlaubsverwaltung.calendar;
 
 import org.junit.jupiter.api.Test;
