+++ src/test/java/org/synyx/urlaubsverwaltung/settings/SettingsWorkingTimeValidatorTest.java
@@ -3,7 +3,7 @@ package org.synyx.urlaubsverwaltung.settings;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.validation.Errors;
-import org.synyx.urlaubsverwaltung.absence.TimeSettings;
+import org.synyx.urlaubsverwaltung.calendar.TimeSettings;
 import org.synyx.urlaubsverwaltung.overtime.OvertimeSettings;
 import org.synyx.urlaubsverwaltung.workingtime.WorkingTimeSettings;
